+++ Sources/CustomEvent/ALKCustomEventCallback.swift
@@ -17,4 +17,5 @@ public protocol ALKCustomEventCallback: AnyObject {
     func conversationCreated(conversationId: String)
     func ratingSubmitted(conversationId: String,rating:Int, comment: String)
     func richMessageClicked(conversationId:String,action:[String:Any], type:String)
+    func conversationInfoClicked()
 }